requirements.txt
# ----------------------------
# Requerimientos para el backend Pokémon Catalog
# ----------------------------

fastapi==0.115.2
uvicorn==0.31.1

# Base de datos
SQLAlchemy==2.0.36

# HTTP y API calls
requests==2.32.3
httpx==0.27.2

# IA real (OpenAI)
openai==1.52.2

# Variables de entorno
python-dotenv==1.0.1
